Fix multiple bugs in fair search
Always put back the value into the threadMVar as soon as possible.
This requires to change the type of the content of the MVar so threads
can see if they are stilled allowed to run.

Remove the stopped threads from the threadMVar. It is safe to assume
that the threads are already in the list of running threads by the time
they want to remove their ids from the list as the MVar is taken before
the thread is started.

Add dummy IORef as a trigger for the finalizer. The naive
implementation, using a weak reference to the Chan does not work as
expected. The problem is the following:
The threads need to be killed when there is no reference on the IOList
anymore. This will be done in the finalizer "killThreads threadVar".
Setting this as a finalizer for the result Chan requires using weak
references to the Chan in the search threads. The GHC runtime now detects
that there is only one reference to the Chan, the reference in
handleResults. Here the program wants to read from the Chan and the GHC
runtime detects that this will be never successful (as there is no
reference to the Chan in another thread). The GHC runtime stops the
waiting thread, the last reference to the Chan is removed, the garbage
collector fires the finalizer and all search threads are killed.