Add SearchQueensLess and remove EditSeq One
Getting one value of EditSeq is trivial, so it is unnecessary to
benchmark it.
SearchQueensLess solves the n queens puzzle with a much smaller
search tree in comparison to SearchQueens. Possibly we can get
fair results by that. Furthermore it is a lot easier to show
the search tree in my thesis.