More realistic EditSeq benchmarks
Benchmark the original EditSeq program given by Michael Hanus as an
example of a real use case. It searches for a minimum value from a
set and then outputs all items equal to this minimal value.
Additionally, this tests which can be parallelized better, simple
searches on a vast search space or programs with a smaller search
space with less fails.

All further EditSeq benchmarks are deleted for the moment.